requirements.txt
streamlit>=1.28.0
pydantic>=2.0
pdfplumber>=0.7.6
python-docx>=0.8.11
Pillow>=9.0.0
pytesseract>=0.3.10
guardrails>=0.6.0
tqdm>=4.0
# LLM & API
langchain>=1.1.0
langchain-openai==1.1.0
langchain-community==0.4.1
openai>=1.0.0

# RAG Pipeline
pypdf>=6.4.1
chromadb>=1.3.5
faiss-cpu>=1.13.1
sentence-transformers>=2.2.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
pylint>=3.0.0